miner/requirements.txt
about-time==4.2.1
alive-progress==3.2.0
certifi==2025.1.31
charset-normalizer==3.4.1
docker==7.1.0
gitdb==4.0.12
GitPython==3.1.44
grapheme==0.6.0
idna==3.10
loguru==0.7.3
numpy==2.2.3
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
six==1.17.0
smmap==5.0.2
tzdata==2025.1
unidiff==0.7.5
urllib3==2.3.0
